requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.9.1
aiosignal==1.3.1
annotated-types==0.7.0
ansible==2.9.6
anticaptchaofficial==1.0.62
anyio==4.5.2
apache-libcloud==2.8.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
argh==0.31.3
async-timeout==5.0.1
attrs==25.1.0
autopep8==2.2.0
bb==0.1
bcrypt==3.1.7
beepy==1.0.7
bitbucket-api==0.5.0
blinker==1.8.2
Brlapi==0.7.0
camera-lib==0.1.0.post202502011633
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.8
cmake==3.26.4
colorama==0.4.3
coloredlogs==15.0.1
contourpy==1.0.7
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dataclasses==0.6
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distro==1.9.0
distro-info==0.23+ubuntu1.1
dnspython==1.16.0
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.2.2
fasteners==0.14.1
filelock==3.12.1
Flask==3.0.3
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.5.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
graphsurgeon==0.4.5
h11==0.16.0
httpcore==1.0.9
httplib2==0.14.0
httpx==0.28.1
humanfriendly==10.0
idna==2.8
importlib-resources==5.12.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.9.1
jmespath==0.9.4
joblib==1.2.0
keyring==18.0.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lit==16.0.5.post0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
MarkupSafe==2.1.3
matplotlib==3.7.1
monotonic==1.5
mpmath==1.3.0
multidict==6.1.0
netaddr==0.7.19
netifaces==0.10.4
networkx==3.1
ntlm-auth==1.1.0
numpy==1.22.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
onnx==1.11.0
onnxruntime==1.15.0
openai==1.86.0
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.72
outcome==1.3.0.post0
p5py==1.0.0
packaging==24.2
pandas==2.0.2
paramiko==2.6.0
pathlib==1.0.1
pep517==0.12.0
pexpect==4.6.0
Pillow==9.5.0
platformdirs==2.5.2
propcache==0.2.0
protobuf==3.20.1
py==1.11.0
pycairo==1.16.2
pycodestyle==2.11.1
pycrypto==2.6.1
pycuda==2021.1
pycups==1.9.73
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.4.0
PyGObject==3.36.0
PyJWT==2.9.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.4
PyPDF2==3.0.1
pyRFC3339==1.1
pyscroll==2.31
PySocks==1.7.1
pytesseract==0.3.13
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
PyTMX==3.32
pytools==2022.1.7
pytz==2023.3
pywinrm==0.3.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.31.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
scikit-learn==1.1.2
scipy==1.9.1
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==2.3.1
selenium==4.27.1
selinux==3.0
sh==2.2.1
simpleaudio==1.0.4
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.12
systemd-python==234
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
trio==0.27.0
trio-websocket==0.11.1
triton==2.0.0
twilio==9.4.4
typing_extensions==4.12.2
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uff==0.6.9
unattended-upgrades==0.1
urllib3==2.2.3
usb-creator==0.3.7
wadllib==1.3.3
websocket-client==1.8.0
Werkzeug==3.0.6
wsproto==1.2.0
xkit==0.0.0
xmltodict==0.12.0
yarl==1.15.2
zipp==3.20.2
